Enables prod and dev mode at the same time for APIs. There are two mqtt brokers, one for the development environment and one for the production environment.
Normal pages decide the environment by the port: users in the port 8000 are in the development environment and users in the port 8080 are in the production environment.

There are two workspaces:

localhost/workspace/dev
and
localhost/workspace/prod

the dev workspace allows to edit the code and develop. Once you like what you see, you just stop the system, run "yarn package" (future "publish button") and start the system again.

yarn package compiles app/next, apps/nextra and apps/api. Compiled versions connect to the production mqtt and listen only for production events. Development versions connect to the development mqtt and listen only for development events.

Removed the port option in the mqtt box in the device definition, since the port depends on the environment. Flashing the device from prod moves the device to prod. Flashing the device from dev, moves the device to dev.

This is a big change, expect some unstabilities in the following days until things settle down.

Co-authored-by: jcarlosn <jcarlos.norte@gmail.com>